+++ CHANGELOG.md
@@ -33,6 +33,7 @@
 #### :bug: Bug fix
 
 - Fix `--create-sourcedirs` generation with for a single project. https://github.com/rescript-lang/rescript/pull/7671
+- Fix rewatch not recompiling on changes under windows. https://github.com/rescript-lang/rescript/pull/7690
 
 # 12.0.0-beta.2